Fix favouriteScreen to FavouriteTemplate
