+++ constants/TabItems.ts
@@ -1,6 +1,6 @@
 import { Icons } from '~/components/atoms/Icons'
 import AccountScreen from '~/components/templates/AccountTemplate'
-import Favorite from '~/components/templates/FavoriteScreen'
+import Favorite from '~/components/templates/FavouriteTempalte'
 import HomeScreen from '~/components/templates/HomeTemplate'
 import NotificationScreen from '~/components/templates/NotificationTemplate'
 import OderScreen from '~/components/templates/OderTemplate'
